+++ rack/config.ru
@@ -1,4 +1,7 @@
 # runs by command: rackup
+
+require_relative 'headers'
+require_relative 'time'
 require_relative 'app'
 
 run App.new
