+++ rack/headers.rb
@@ -1,0 +1,5 @@
+# require 'headers'
+
+class Headers
+  HEADERS = { 'Content-Type' => 'text/plain' }.freeze
+end
